prevent empty vfilter
The current code adds a trailing comma in case no extra video filters
are defined. At least with my version of ffmpeg (5.1.6-0+deb12u1) this
does not work and leads to:

No such filter: ''

Which is: ffmpeg interprets the (empty) part after the comma as filter
and does not have such a filter.

The patch only adds the comma if the extra filters are defined and are
not empty.